requirements.txt
fastapi==0.104.0
qdrant-client==1.6.4
langchain==0.0.325
numpy==1.26.1
pandas==2.1.1
uvicorn==0.23.2
sentence-transformers==2.2.2
sentencepiece==0.1.99
python-dotenv==1.0.0
